requirements.txt
absl-py==0.12.0
astor==0.8.1
astroid==2.8.3
astunparse==1.6.3
bokeh==2.3.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
Flask==2.0.0
flatbuffers==1.12
future==0.18.2
gast==0.2.2
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.17
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.0.1
imutils==0.5.4
isort==5.9.3
itsdangerous==2.0.0
Jinja2==3.0.0
joblib==1.0.1
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.0
matplotlib==3.4.1
mccabe==0.6.1
networkx==2.5.1
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python==4.5.2.52
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==21.0
pandas==1.2.4
Pillow==8.2.0
platformdirs==2.4.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.11.1
pyparsing==2.4.7
pyrealsense2==2.44.0.3073
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.2
scikit-plot==0.3.7
scipy==1.6.3
seaborn==0.11.1
six==1.15.0
sklearn==0.0
tensorboard==2.1.1
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.0.3
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.2
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2021.6.6
toml==0.10.2
torch==1.7.0+cu101
torchvision==0.8.1+cu101
tornado==6.1
tqdm==4.60.0
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.4
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
